+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- [#113](https://github.com/XenitAB/spegel/pull/113) Remove image filter configuration.
+
 ### Fixed
 
 ### Security